backend/requirements.txt
#requirement.txt
fastapi==0.101.0
uvicorn[standard]==0.23.0
sqlmodel==0.0.8
SQLAlchemy==2.1.0
python-multipart==0.0.6
pillow==10.0.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
requests==2.31.0
python-dotenv==1.0.0
opencv-python==4.9.0.96
torch==2.2.0
